Omero part fixed
