+++ src/KOMP_Images_ETL.egg-info/SOURCES.txt
@@ -1,0 +1,14 @@
+README.md
+pyproject.toml
+setup.cfg
+setup.py
+src/download_from_drive.py
+src/download_from_omero.py
+src/resubmit.py
+src/utils.py
+src/KOMP_Images_ETL.egg-info/PKG-INFO
+src/KOMP_Images_ETL.egg-info/SOURCES.txt
+src/KOMP_Images_ETL.egg-info/dependency_links.txt
+src/KOMP_Images_ETL.egg-info/entry_points.txt
+src/KOMP_Images_ETL.egg-info/requires.txt
+src/KOMP_Images_ETL.egg-info/top_level.txt
