+++ src/KOMP_Images_ETL.egg-info/requires.txt
@@ -1,0 +1,9 @@
+requests
+paramiko
+mysql.connector
+
+[:python_version == "3.9"]
+importlib-metadata
+
+[:python_version >= "3.9"]
+importlib-metadata
