+++ src/download_from_drive.py
@@ -1,12 +1,10 @@
 import collections
-import datetime
-import logging
+from datetime import datetime
 import shutil
 from typing import Any
 
 import mysql.connector
 import os
-import sys
 import utils
 import paramiko
 
@@ -44,30 +42,31 @@ class image_upload_status(object):
         self.Message = Message
 
 
-def update_images_status(record: dict,
-                         imagefilekey: int):
-    if not record:
+def update_images_status(imageDict: dict, imagefilekey):
+    if not imageDict:
         raise ValueError("Nothing to be inserted")
 
-    conn = db_init(server=db_server, username=db_username, password=db_password, database=db_name)
-    cursor = conn.cursor()
+    db_server = "rslims.jax.org"
+    db_user = "dba"
+    db_password = "rsdba"
+    db_name = "komp"
+    conn = db_init(server=db_server, username=db_user, password=db_password, database=db_name)
+    cursor1 = conn.cursor()
 
-    '''Remove duplicate records'''
-
+    '''Remove deplciate records'''
     cleanStmt = """ DELETE i FROM komp.imagefileuploadstatus i, komp.imagefileuploadstatus j 
                     WHERE i._ImageFile_key > j._ImageFile_key AND i.SourceFileName = j.SourceFileName; 
                 """
-    cursor.execute(cleanStmt)
 
-    # placeholders = ', '.join(['%s'] * len(record))
-    columns = ', '.join(record.keys())
-    logger.debug(f"Columns are {columns}")
+    cursor1.execute(cleanStmt)
+    conn.commit()
 
+    cursor2 = conn.cursor()
     sql = "UPDATE KOMP.imagefileuploadstatus SET {} WHERE _ImageFile_key = {};".format(
-        ', '.join("{}='{}'".format(k, v) for k, v in record.items()), imagefilekey)
-    print(sql)
+        ', '.join("{}='{}'".format(k, v) for k, v in imageDict.items()), imagefilekey)
     logger.debug(sql)
-    cursor.execute(sql, list(record.values()))
+    print(sql)
+    cursor2.execute(sql)
     conn.commit()
     conn.close()
 
@@ -98,9 +97,9 @@ def generate_file_location(conn: mysql.connector.connection,
 
         IMPC_Code = record["DestinationFileName"].split("/")[4]
         temp = record["SourceFileName"].split("\\")[4:]
-        drive_path = "/Volumes/"  # If you are on mac/linux
-        fileLocation = os.path.join(drive_path, *temp)
-        # fileLocation = "//" + os.path.join("bht2stor.jax.org\\", *temp).replace("\\", "/") #If you are on windows
+        # drive_path = "/Volumes/"  # If you are on mac/linux
+        # fileLocation = os.path.join(drive_path, *temp)
+        fileLocation = "//" + os.path.join("bht2stor.jax.org\\", *temp).replace("\\", "/")  # If you are on windows
         logger.debug(f"Source file path is {fileLocation}")
 
         fileLocationMap[IMPC_Code].append([int(record["_ImageFile_key"]), fileLocation])
@@ -158,7 +157,7 @@ def download_from_drive(fileLocationDict: collections.defaultdict[list],
                 logger.error(e)
 
                 """Create object"""
-                file_Status = image_upload_status(DateOfUpload=datetime.datetime.now(),
+                file_Status = image_upload_status(DateOfUpload=datetime.today().strftime('%Y-%m-%d'),
                                                   UploadStatus="Fail",
                                                   Message="File not found on the disk")
 
@@ -184,7 +183,7 @@ def send_to_server(file_to_send: str,
         try:
             logger.info(ftp_client.stat('/pictures/' + IMPC_Code + "/" + file_to_send))
             logger.info(f'File exists in directory {IMPC_Code}')
-            file_Status = image_upload_status(DateOfUpload=datetime.datetime.now(),
+            file_Status = image_upload_status(DateOfUpload=datetime.today().strftime('%Y-%m-%d'),
                                               UploadStatus="Success",
                                               Message="File already exits on server")
 
@@ -195,7 +194,7 @@ def send_to_server(file_to_send: str,
             ftp_client.put(download_to + "/" + file_to_send,
                            "pictures/" + IMPC_Code + "/" + file_to_send)
 
-            file_Status = image_upload_status(DateOfUpload=datetime.datetime.now(),
+            file_Status = image_upload_status(DateOfUpload=datetime.today().strftime('%Y-%m-%d'),
                                               UploadStatus="Success",
                                               Message="File successfully uploaded to server")
 
@@ -203,6 +202,7 @@ def send_to_server(file_to_send: str,
 
         # os.remove(os.path.join(download_to, loc[1].split("/")[-1]))
         ftp_client.close()
+        logger.info(f"Finish uploading {file_to_send}")
 
     except paramiko.SSHException:
         logger.error("Connection Error")
@@ -240,14 +240,20 @@ def main():
 
 # Press the green button in the gutter to run the script.
 if __name__ == '__main__':
-    logger = logging.getLogger("__main__")
+
+    job_name = 'download_from_drive'
+    logging_dest = os.path.join(utils.get_project_root(), "logs")
+    date = datetime.now().strftime("%B-%d-%Y")
+    logging_filename = logging_dest + "/" + f'{date}.log'
+    logger = utils.createLogHandler(job_name, logging_filename)
+    logger.info('Logger has been created')
 
     db_username = utils.db_username
     db_password = utils.db_password
     db_server = utils.db_server
     db_name = utils.db_name
 
-    download_to = "/Users/chent/Desktop/Pictures"
+    download_to = os.path.join(utils.get_project_root(), "Pictures")
     try:
         os.mkdir(download_to)
